move related text methods to wynntilsscreen
